+++ app/controllers/adverse_events_controller.rb
@@ -25,29 +25,9 @@ class AdverseEventsController < ApplicationController
 
   # GET /adverse-events
   def index
-    adverse_event_scope = viewable_adverse_events.search(params[:search])
-    adverse_event_scope = adverse_event_scope.with_site(params[:site_id]) if params[:site_id].present?
-    adverse_event_scope = adverse_event_scope.where(user_id: params[:reported_by_id]) if params[:reported_by_id].present?
-    adverse_event_scope = adverse_event_scope.where(closed: params[:status] == 'closed') if params[:status].present?
-    @order = params[:order]
-    case params[:order]
-    when 'adverse_events.reported_by'
-      adverse_event_scope = adverse_event_scope.includes(:user).order('users.last_name, users.first_name')
-    when 'adverse_events.reported_by desc'
-      adverse_event_scope = adverse_event_scope.includes(:user).order('users.last_name desc, users.first_name desc')
-    when 'adverse_events.site_name'
-      adverse_event_scope = adverse_event_scope.includes(subject: :site).order('sites.name')
-    when 'adverse_events.site_name desc'
-      adverse_event_scope = adverse_event_scope.includes(subject: :site).order('sites.name desc')
-    when 'adverse_events.subject_code'
-      adverse_event_scope = adverse_event_scope.includes(:subject).order('subjects.subject_code')
-    when 'adverse_events.subject_code desc'
-      adverse_event_scope = adverse_event_scope.includes(:subject).order('subjects.subject_code desc')
-    else
-      @order = scrub_order(AdverseEvent, params[:order], 'adverse_events.created_at desc')
-      adverse_event_scope = adverse_event_scope.order(@order)
-    end
-    @adverse_events = adverse_event_scope.page(params[:page]).per(40)
+    scope = viewable_adverse_events.search(params[:search]).includes(:subject, { subject: :site }, :user)
+    scope = filter_adverse_events(scope)
+    @adverse_events = order_adverse_events(scope).page(params[:page]).per(40)
   end
 
   # # GET /adverse-events/1
@@ -69,6 +49,7 @@ class AdverseEventsController < ApplicationController
     if @adverse_event.save
       @adverse_event.create_notifications
       @adverse_event.send_email_in_background
+      @adverse_event.generate_number!
       redirect_to [@project, @adverse_event], notice: 'Adverse event was successfully created.'
     else
       render :new
@@ -128,5 +109,17 @@ class AdverseEventsController < ApplicationController
       # Attribute Accessor
       :subject_code, :event_date
     )
+  end
+
+  def filter_adverse_events(scope)
+    scope = scope.with_site(params[:site_id]) if params[:site_id].present?
+    scope = scope.where(user_id: params[:reported_by_id]) if params[:reported_by_id].present?
+    scope = scope.where(closed: params[:status] == 'closed') if params[:status].present?
+    scope
+  end
+
+  def order_adverse_events(scope)
+    @order = params[:order]
+    scope.order(AdverseEvent::ORDERS[params[:order]] || AdverseEvent::DEFAULT_ORDER)
   end
 end