+++ app/models/subject.rb
@@ -16,8 +16,8 @@ class Subject < ApplicationRecord
   scope :with_missing_design_on_event, -> (design, event) { where('subjects.id IN (select subject_events.subject_id from subject_events where subject_events.event_id = ? and subject_events.id IN (SELECT sheets.subject_event_id from sheets where sheets.deleted = ? and sheets.missing = ? and sheets.design_id = ? and sheets.subject_event_id IS NOT NULL))', event, false, true, design) }
   scope :with_unentered_design_on_event, -> (design, event) { where('subjects.id IN (select subject_events.subject_id from subject_events where subject_events.event_id = ? and subject_events.id NOT IN (SELECT sheets.subject_event_id from sheets where sheets.deleted = ? and sheets.design_id = ? and sheets.subject_event_id IS NOT NULL))', event, false, design) }
   scope :without_design_on_event, -> (design, event) { where('subjects.id NOT IN (select subject_events.subject_id from subject_events where subject_events.event_id = ? and subject_events.id IN (SELECT sheets.subject_event_id from sheets where sheets.deleted = ? and sheets.design_id = ?))', event, false, design) }
-  scope :randomized, -> { joins(:randomizations).distinct }
-  scope :unrandomized, -> { joins('LEFT OUTER JOIN randomizations ON randomizations.subject_id = subjects.id and randomizations.deleted is false').where('randomizations.id IS NULL').distinct }
+  scope :randomized, -> { where.not(randomizations_count: 0) }
+  scope :unrandomized, -> { where(randomizations_count: 0) }
   scope :open_aes, -> { joins(:adverse_events).where(adverse_events: { closed: false }).distinct }
   scope :closed_aes, -> { joins(:adverse_events).where(adverse_events: { closed: true }).distinct }
   scope :any_aes, -> { joins(:adverse_events).distinct }