+++ db/migrate/20170224203450_add_variables_count_to_domains.rb
@@ -1,0 +1,5 @@
+class AddVariablesCountToDomains < ActiveRecord::Migration[5.0]
+  def change
+    add_column :domains, :variables_count, :integer, null: false, default: 0
+  end
+end