+++ db/migrate/20170227205756_add_randomizations_count_to_subjects.rb
@@ -1,0 +1,6 @@
+class AddRandomizationsCountToSubjects < ActiveRecord::Migration[5.0]
+  def change
+    add_column :subjects, :randomizations_count, :integer, null: false, default: 0
+    add_index :subjects, :randomizations_count
+  end
+end