+++ CHANGELOG.md
@@ -7,11 +7,15 @@
 ## 0.50.0
 
 ### Enhancements
+- **General Changes**
+  - Optimized loading times for a number of index pages
 - **Gem Changes**
   - Updated to highcharts 5.0.7
 - **Design Changes**
   - Scale headers now stick to the top of the screen when scrolling past a set
     of scale variables
+- **Domain Changes**
+  - Domains can now be sorted by number of variables
 - **Export Changes**
   - Adverse Event IDs are now exported as numbers
 - **Handoff Changes**
@@ -24,6 +28,7 @@
       - This change helps mark handoff designs complete that only include
         sections with descriptions
 - **Gem Changes**
+  - Updated to rails 5.0.2.rc1
   - Updated to clipboard.js 1.6.0
 
 ### Bug Fixes